V1.3
Individual Voice Configs to allow better support for multiple agents. Was either this or a very long list in the config file. This probably makes it easier to manage in the long run.
Idea is we can now call a method like "set narrator adam" to dynamically load a config file "adam.voice.json" or reload on the fly for tweaking settings without reloading all voices.